src/main/java/nextstep/subway/line/domain/Stations.java
package nextstep.subway.line.domain;

import nextstep.subway.station.domain.Station;
import nextstep.subway.station.dto.StationResponse;

import java.util.List;
import java.util.stream.Collectors;

public class Stations {
    private final List<Station> stations;

    public Stations(List<Station> stations) {
        this.stations = stations;
    }

    public List<StationResponse> toResponses() {
        return stations.stream()
                .map(StationResponse::of)
                .collect(Collectors.toList());
    }
}
